+++ tests/test_bpy_handlers_rst.py
@@ -1,0 +1,61 @@
+#!/usr/bin/env python
+# Copyright (C) 2013 by Yu-Jie Lin
+# 
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+# 
+# The above copyright notice and this permission notice shall be included in
+# all copies or substantial portions of the Software.
+# 
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+# THE SOFTWARE.
+
+
+import unittest
+import doctest
+
+from bpy.handlers.rst import Handler
+
+
+class HandlerTestCase(unittest.TestCase):
+
+  def setUp(self):
+
+    self.handler = Handler(None)
+
+  def tearDown(self):
+
+    self.handler = None
+
+  def test_id_affix(self):
+
+    handler = self.handler
+    handler.title = 'test'
+    source = '''Test Handler
+------------'''
+
+    html_base = '<div class="section" id="%stest-handler">\n<h2>Test Handler</h2>\n</div>\n'
+
+    html = html_base % ''
+    self.assertEqual(handler.generate(source), html)
+
+    handler.header['id_affix'] = ''
+    html = html_base % '098f-'
+    self.assertEqual(handler.generate(source), html)
+    self.assertEqual(handler.modified, True)
+    self.assertEqual(handler.generate_header(), '''.. !b
+   id_affix: 098f
+''')
+
+    handler.header['id_affix'] = 'foobar-prefix'
+    html = html_base % 'foobar-prefix-'
+    self.assertEqual(handler.generate(source), html)